requirements.txt
Django>=5.2
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
rabbitmq-server
python-dotenv
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
gunicorn>=20.1
